feat: create analysis controller

+++ api/liftform/src/main/java/com/rianlucassb/liftform/core/usecases/analysis/create/CreateAnalysisUseCaseOutput.java
@@ -3,6 +3,6 @@ package com.rianlucassb.liftform.core.usecases.analysis.create;
 public record CreateAnalysisUseCaseOutput(
         Long analysisId,
         String uploadUrl,
-        int expiresIn
+        Long expiresIn
 ) {
 }
